requirements.txt
# requirements.txt
fastapi
uvicorn
openai>=1.0.0
python-dotenv
pydantic
pytest
black
mypy
pdfplumber
python-docx
requests
matplotlib